Re-add support for COG rendering using maplibre plugin
This adds back support for rendering COGs in Web Mercator using
the Maplibre COG protocol plugin, since the deck.gl-based solution
isn't compatible with our build targets right now.

See https://github.com/OpenGeoMetadata/ogm-viewer/issues/100.
