+++ package.json
@@ -59,6 +59,7 @@
     "@deck.gl/mapbox": "^9.3.1",
     "@deck.gl/mesh-layers": "^9.3.1",
     "@developmentseed/deck.gl-geotiff": "^0.7.0",
+    "@geomatico/maplibre-cog-protocol": "^0.9.0",
     "@iiif/presentation-2": "^1.0.4",
     "@iiif/presentation-3": "^2.2.3",
     "@luma.gl/core": "^9.3.3",
